% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataFrames.r
\name{compareDataFrames}
\alias{compareDataFrames}
\alias{dfcomp}
\title{Compare two data frames}
\usage{
compareDataFrames(old, new, rowKeys, sort.rows.first = TRUE,
  keepRowTag = FALSE, row.tag = "my.row.tag_", verbose = TRUE)
}
\arguments{
\item{old}{first   data.frame to be compared.}

\item{new}{second data.frame to be compared.}

\item{rowKeys}{a character vector with the column names which identify unique rows.}

\item{sort.rows.first}{see details}

\item{keepRowTag}{see details}

\item{row.tag}{see details}

\item{verbose}{verbose}
}
\value{
An S3 object with class "dfcomp" (data.frame comparison).
Basically a list with the following items (all of them data.frames):
"del.cols"
"new.cols"
"dif.col.class"
"dup.rows.old"
"dup.rows.new"
"del.rows"
"new.rows"
"dif.cells"
}
\description{
The function compares two data.frames.
}
\details{
Details...
}
