% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataFrames.r
\name{print.dfcomp}
\alias{print.dfcomp}
\title{print method for a dfcomp list}
\usage{
\method{print}{dfcomp}(x, report.dif.col.class = FALSE)
}
\arguments{
\item{x}{a "dfcomp" object created by the compareDataFrames function.}

\item{report.dif.col.class}{if TRUE differences in the class of the columns of the data frames are reported.}
}
\description{
print method for a dfcomp list
}
\seealso{
compareDataFrames
}
