% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataFrames.r
\name{summary.dfcomp}
\alias{summary.dfcomp}
\title{summary method for a dfcomp list}
\usage{
\method{summary}{dfcomp}(x, verbose = TRUE, report.dif.col.class = TRUE)
}
\arguments{
\item{x}{a dfcomp element created by compareDataFrames.}

\item{verbose}{verbose.}

\item{report.dif.col.class}{if TRUE differences in the class of the columns of the data frames are reported.}
}
\description{
summary method for a dfcomp list
}
\seealso{
compareDataFrames
}
